refactor: moves and improves MainActivity.kt
